refactor: util함수 분리
