+++ src/components/molecules/screens/library/LibrarySeatingChart.tsx
@@ -7,54 +7,16 @@ import {
 } from '../../../../configs/utility/librarySeatingChart';
 import {RoomNameType} from '../../../../configs/utility/librarySeatingChart/roomName';
 import SeatItem from '../../../atoms/seat_item/SeatItem';
-import {SeatStatusType} from '../../../../configs/utility/librarySeatingChart/seatStatus';
 import {
-  ROOM_2_FOR_DISABLED_PERSON_LIST,
-  ROOM_4_FOR_DISABLED_PERSON_LIST,
-} from '../../../../configs/utility/librarySeatingChart/forDisabledPersonList';
+  findForDisabledPerson,
+  findSeatStatusBySeatId,
+} from '../../../../utils/library/findSeatItem';
 
 const {width} = Dimensions.get('screen');
 
 type Props = {
   roomNumber: RoomNameType;
   handlePressItem: () => void;
-};
-
-const mock: Array<{seatId: number; status: SeatStatusType}> = [
-  {
-    seatId: 12,
-    status: 'NOT_AVAILABLE',
-  },
-  {
-    seatId: 14,
-    status: 'RESERVED',
-  },
-  {
-    seatId: 23,
-    status: 'SPECIFIED',
-  },
-];
-
-const findStatusBySeatId = (seatId: number) => {
-  return mock.find(item => item.seatId === seatId)?.status;
-};
-
-const findForDisabledPerson = (
-  roomNumber: Props['roomNumber'],
-  seatId: number,
-): boolean | null => {
-  switch (roomNumber) {
-    case '2':
-      return (
-        ROOM_2_FOR_DISABLED_PERSON_LIST.some(item => item === seatId) ?? null
-      );
-    case '4':
-      return (
-        ROOM_4_FOR_DISABLED_PERSON_LIST.some(item => item === seatId) ?? null
-      );
-    default:
-      return null;
-  }
 };
 
 const LibrarySeatingChart = ({roomNumber, handlePressItem}: Props) => {
@@ -90,7 +52,7 @@ const LibrarySeatingChart = ({roomNumber, handlePressItem}: Props) => {
               scrollEnabled={false}
               data={rowItem}
               renderItem={({item}) => {
-                const status = findStatusBySeatId(item);
+                const status = findSeatStatusBySeatId(item);
                 const forDisabledPerson = findForDisabledPerson(
                   roomNumber,
                   item,
